BasketApi
Added BasketApi - code to connect to the database and a simple route to add a document has been created but not tested to see if it works yet.
Added the ability to check whether the active date in the offerings data is before the current date when sending the product offerings to the frontend.
Added a project template for the CheckoutApi.