BasketController
in the route to add a new item to the basket:
now inserts new items at the beginning of the list instead of the end
will now check if an item already exists or not, if it does exist it will add the quantities of the 2 offerings together.
Since we can also change the quantity of items in the view basket page we will need to a different route for that where we are only updating the items and not attempting to add new ones.