BasketDiscController - Route to add items
Not fully tested but for the most part it seems to work correctly.
Simplified the requests that the frontend will need to make, only need to send the offering_key and the quantity the user is wanting to add to their cart.
Using interprocess communication to get the offering from the catalog microservice and the tiers when its a BULK_DISCOUNT.
Handles BULK_DISCOUNTS and checks the MaxQty on everything.